fixed invalid user updating when trying to update hidden user hopefully fixed executing on Win9x due to missing DLL export small memory save in Waiting users queue fixed some old links to berlios

+++ trunk/client/Socket.cpp
@@ -58,7 +58,7 @@ string SocketException::errorToString(int aError) throw() {
 	return msg;
 }
 
-void Socket::create(int aType /* = TYPE_TCP */) throw(SocketException) {
+void Socket::create(uint8_t aType /* = TYPE_TCP */) throw(SocketException) {
 	if(sock != INVALID_SOCKET)
 		disconnect();
